Separate out query functions into a separate class.
My doing all of the query work in a separate class instance,
we can pass that class instance to the nodes so they will
be able to make queries when the user invokes the query command.

Pass the Query class object to the device nodes and enable
the query command to call the various querys.