If node not found add it.
It seems like the balance and totalusage channels don't show up
in the intitial device info query but usage information for them
is present in the usage query data.

So if we see it in the usage, but it hasn't yet been created,
create it so that we can start tracking usage.