Always call setDriver with force = True
On every poll/query, send the data even if it hasn't changed.

Historically, we tried to limit the amount of data sent by the
node server to only changed data so we didn't overload the ISY/
IoX.  But we seem to be seeing more issues with the IoX not
getting the initial data and then never getting updated because
some of the data changes infrequently.